Move TRole type to roles.ts
