+++ front/app/services/permissions/permissions.ts
@@ -11,21 +11,6 @@ import useAuthUser from 'hooks/useAuthUser';
 import useAppConfiguration from 'hooks/useAppConfiguration';
 import { isNilOrError } from 'utils/helperUtils';
 
-export interface IProjectModeratorRole {
-  type: 'project_moderator';
-  project_id: string;
-}
-
-interface IAdminRole {
-  type: 'admin';
-}
-
-export interface IRoleRegisty {
-  IAdminRole: IAdminRole;
-  IProjectModeratorRole: IProjectModeratorRole;
-}
-
-export type TRole = IRoleRegisty[keyof IRoleRegisty];
 export interface IRouteItem {
   type: 'route';
   path: string;
